Add empty conversations list
refs: MBL-11245
affects: Parent
release note: none
