+++ Core/CoreTests/API/APIConversationTests.swift
@@ -24,4 +24,53 @@ class APIConversationTests: CoreTestCase {
     func testMake() {
         XCTAssertEqual(APIConversation.make(messages: [.make()]).messages?.count, 1)
     }
+
+    func testGetConversationsUnreadCountRequest() {
+        XCTAssertEqual(GetConversationsUnreadCountRequest().path, "conversations/unread_count")
+        let response = try? JSONDecoder().decode(GetConversationsUnreadCountRequest.Response.self, from: """
+            { "unread_count": "5" }
+            """.data(using: .utf8)!)
+        XCTAssertEqual(response?.unread_count, 5)
+    }
+
+    func testGetConversationsRequest() {
+        XCTAssertEqual(GetConversationsRequest().path, "conversations")
+        XCTAssertEqual(GetConversationsRequest().queryItems, [])
+        XCTAssertEqual(GetConversationsRequest(include: [.participant_avatars], perPage: 50, scope: .sent).queryItems, [
+            URLQueryItem(name: "include[]", value: "participant_avatars"),
+            URLQueryItem(name: "per_page", value: "50"),
+            URLQueryItem(name: "scope", value: "sent"),
+        ])
+    }
+
+    func testGetConversationRequest() {
+        XCTAssertEqual(GetConversationRequest(id: "1").path, "conversations/1")
+        XCTAssertEqual(GetConversationRequest(id: "1").queryItems, [])
+        XCTAssertEqual(GetConversationRequest(id: "2", include: [.participant_avatars]).queryItems, [
+            URLQueryItem(name: "include[]", value: "participant_avatars"),
+        ])
+    }
+
+    func testPutConversationRequest() {
+        let request = PutConversationRequest(id: "1", workflowState: .read)
+        XCTAssertEqual(request.path, "conversations/1")
+        XCTAssertEqual(request.method, .put)
+        XCTAssertEqual(request.body?.id, "1")
+        XCTAssertEqual(request.body?.workflow_state, .read)
+    }
+
+    func testPostAddMessageRequest() throws {
+        let request = PostAddMessageRequest(id: "1", message: .init(
+            recipients: ["1"],
+            body: "This is a reply",
+            subject: "Subject One",
+            attachment_ids: nil,
+            media_comment_id: nil,
+            media_comment_type: nil,
+            context_code: nil,
+            bulk_message: nil)
+        )
+        XCTAssertEqual(request.path, "conversations/1/add_message")
+        XCTAssertEqual(request.method, .post)
+    }
 }
